better source import paths
