+++ src/fdtdx/objects/sources/__init__.py
@@ -12,3 +12,21 @@ Key constraints include:
 The constraints are implemented as modules that can be chained together
 and applied during the optimization process.
 """
+
+from .plane_source import (
+    LinearlyPolarizedPlaneSource, 
+    GaussianPlaneSource, 
+    ConstantAmplitudePlaneSource,
+    HardConstantAmplitudePlanceSource,
+)
+
+from .profile import SingleFrequencyProfile, GaussianPulseProfile
+
+__all__ = [
+    "LinearlyPolarizedPlaneSource",
+    "GaussianPlaneSource", 
+    "ConstantAmplitudePlaneSource",
+    "HardConstantAmplitudePlanceSource",
+    "SingleFrequencyProfile",
+    "GaussianPulseProfile",
+]
